Updat econtrol channel usage to be more idiomatic and use struct{}
This updates the code to use a closing channel to signal when it's time for
things to shut down, instead of an unused value. This is more idiomatic in Go.

If we were using the value being sent through the channel, then we would not
make this change.

The reason we use `struct{}` is because it's nothing, creating an empty struct
incurs no memory allocation.

Signed-off-by: Tim Heckman <t@heckman.io>