Update validateRequest() to use error values for returns
This is much more idiomatic in that we're using an `error` value to indicate an
error, versus combining boolean and string values together to figure out what
the error condition is.

There are even better ways of propagating errors and their meanings, but they
are a bit more verbose and would likely make it harder to understand.

Signed-off-by: Tim Heckman <t@heckman.io>